Adjust assignment deletion to be consistent with Realm sync (adjust for errors with try Realm)

+++ Studium/Controllers/List Controllers/AssignmentsViewController.swift
@@ -152,15 +152,20 @@ class AssignmentsViewController: SwipeTableViewController, UISearchBarDelegate, 
     override func updateModelDelete(at indexPath: IndexPath) {
         let cell = tableView.cellForRow(at: indexPath) as! DeletableEventCell
             do{
-                try self.realm.write{
-                    if let assignment = cell.event as? Assignment{
-                        let parentCourse = assignment.parentCourse[0]
-                        let assignmentIndex = parentCourse.assignments.index(of: assignment)
-                        parentCourse.assignments.remove(at: assignmentIndex!)
-                        assignment.deleteNotifications()
+                if let user = app.currentUser {
+                    realm = try! Realm(configuration: user.configuration(partitionValue: user.id))
+                    try self.realm.write{
+                        if let assignment = cell.event as? Assignment{
+                            let parentCourse = assignment.parentCourse[0]
+                            let assignmentIndex = parentCourse.assignments.index(of: assignment)
+                            parentCourse.assignments.remove(at: assignmentIndex!)
+                            assignment.deleteNotifications()
+                        }
+                        cell.event!.deleteNotifications()
+                        self.realm.delete(cell.event!)
                     }
-                    cell.event!.deleteNotifications()
-                    self.realm.delete(cell.event!)
+                }else{
+                    print("error accessing user")
                 }
             }catch{
                 print("Error deleting OtherEvent")
